refactor: DTO 패키지 재구성 — domain/ 혼재 DTO를 dto/ 하위로 통합
- domain/{category,comment,member,file} DTO → dto/{category,comment,auth,file} 이동
- MemberLoginReqDto → dto/auth/LoginReqDto 이동 및 클래스명 변경
- 미사용 TokenReqDto 삭제
- wildcard import(domain.comment.*) → 명시적 import 분리
