--- backend/src/main/java/com/moya/myblogboot/domain/token/TokenReqDto.java
@@ -1,11 +1,0 @@
-package com.moya.myblogboot.domain.token;
-
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenReqDto {
-    private Long refreshTokenKey;
-}
