+++ backend/src/main/java/com/moya/myblogboot/dto/auth/LoginReqDto.java
@@ -1,4 +1,4 @@
-package com.moya.myblogboot.domain.member;
+package com.moya.myblogboot.dto.auth;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -8,7 +8,7 @@ import lombok.*;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class MemberLoginReqDto {
+public class LoginReqDto {
     @NotBlank(message = "아이디를 입력하세요.")
     @Size(min = 6, max = 20, message = "아이디는 6자 이상 20자 이하로 입력하여야 합니다.\n※ 공백 및 특수문자 입력 불가능.")
     private String username;
