+++ backend/src/main/java/com/moya/myblogboot/dto/category/CategoriesResDto.java
@@ -1,4 +1,6 @@
-package com.moya.myblogboot.domain.category;
+package com.moya.myblogboot.dto.category;
+
+import com.moya.myblogboot.domain.category.Category;
 
 
 import lombok.*;
