+++ backend/src/main/java/com/moya/myblogboot/repository/CategoryQuerydslRepository.java
@@ -1,6 +1,6 @@
 package com.moya.myblogboot.repository;
 
-import com.moya.myblogboot.domain.category.CategoriesResDto;
+import com.moya.myblogboot.dto.category.CategoriesResDto;
 
 import java.util.List;
