+++ backend/src/main/java/com/moya/myblogboot/service/AuthService.java
@@ -1,12 +1,12 @@
 package com.moya.myblogboot.service;
 
-import com.moya.myblogboot.domain.member.MemberLoginReqDto;
+import com.moya.myblogboot.dto.auth.LoginReqDto;
 import com.moya.myblogboot.domain.token.Token;
 import com.moya.myblogboot.domain.token.TokenInfo;
 
 public interface AuthService {
 
-    Token adminLogin(MemberLoginReqDto loginReqDto);
+    Token adminLogin(LoginReqDto loginReqDto);
 
     String reissuingAccessToken(String refreshToken);
